Finishing touches to automod config
Also moved automod to mod cog instead of it's own cog
