+++ utils/exceptions.py
@@ -13,3 +13,13 @@ class InvalidUrl(commands.BadArgument):
 class InvalidAutomodModule(commands.BadArgument):
     def __init__(self, module: str):
         self.module = module
+
+
+class AutomodModuleNotEnabled(commands.BadArgument):
+    def __init__(self, module: str):
+        self.module = module
+
+
+class AutomodModuleAlreadyEnabled(commands.BadArgument):
+    def __init__(self, module: str):
+        self.module = module
